Prof s jonalista

+++ DoCeluNaCzasWebApi/Services/UpdateService/Helpers/UpdateServiceHelper.cs
@@ -1,8 +1,10 @@
-﻿using DCNC.Bussiness.PublicTransport.JsonData;
+﻿using DCNC.Bussiness.PublicTransport.JoiningTrips;
+using DCNC.Bussiness.PublicTransport.JsonData;
 using DCNC.Bussiness.PublicTransport.JsonData.General;
 using DCNC.DataAccess.PublicTransport.Helpers;
 using DCNC.Service.Caching;
 using DCNC.Service.Caching.Helpers;
+using DCNC.Service.Database.Interfaces;
 using DCNC.Service.PublicTransport.JsonData.Abstracts.Interfaces;
 using DCNC.Service.PublicTransport.Time;
 using DoCeluNaCzasWebApi.Services.Delays;
@@ -12,8 +14,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using DCNC.Bussiness.PublicTransport.JoiningTrips;
-using DCNC.Service.Database.Interfaces;
 
 // ReSharper disable PossibleNullReferenceException
 namespace DoCeluNaCzasWebApi.Services.UpdateService.Helpers
@@ -32,8 +32,8 @@ namespace DoCeluNaCzasWebApi.Services.UpdateService.Helpers
         readonly BusStopModelService _busStopModelService;
         readonly IDocumentStoreRepository _documentStoreRepository;
 
-        public UpdateServiceHelper(Joiner joiner, Grouper grouper, TimeService timeService, IJsonDataService tripService, 
-                                   IJsonDataService busStopService, IJsonDataService busLineService, IJsonDataService expeditionService, 
+        public UpdateServiceHelper(Joiner joiner, Grouper grouper, TimeService timeService, IJsonDataService tripService,
+                                   IJsonDataService busStopService, IJsonDataService busLineService, IJsonDataService expeditionService,
                                    IJsonDataService stopInTripService, BusStopModelService busStopModelService, IDocumentStoreRepository documentStoreRepository)
         {
             _joiner = joiner;
@@ -56,7 +56,8 @@ namespace DoCeluNaCzasWebApi.Services.UpdateService.Helpers
             var expeditionsAsJObject = await _expeditionService.GetDataAsJObjectAsync(Urls.Expedition, JsonType.Expedition);
             var stopsInTripsAsJObject = await _stopInTripService.GetDataAsJObjectAsync(Urls.StopsInTrips, JsonType.StopInTrip);
 
-            return (tripsAsJObject, busStopsAsJObject, busLinesAsJObject, expeditionsAsJObject, stopsInTripsAsJObject);
+            return (tripsAsJObject, busStopsAsJObject, busLinesAsJObject, expeditionsAsJObject,
+                stopsInTripsAsJObject);
         }
 
         public void SetAndCache(JObject tripsAsJObject, JObject busStopsAsJObject, JObject busLinesAsJObject, JObject expeditionsAsJObject, JObject stopsInTripsAsJObject)
@@ -68,7 +69,7 @@ namespace DoCeluNaCzasWebApi.Services.UpdateService.Helpers
             var expeditionData = _expeditionService.GetData<ExpeditionData>(expeditionsAsJObject).FirstOrDefault();
 
             var busStopDataModel = _busStopModelService.JoinBusStopData(busStopDataList);
-            var tripsWithBusStops = _joiner.GetTripsWithBusStopList(tripDataList, busStopDataList, busLineDataList, stopInTripDataList, expeditionData);// to do bazy
+            var tripsWithBusStops = _joiner.GetTripsWithBusStopList(tripDataList, busStopDataList, busLineDataList, stopInTripDataList, expeditionData);
 
             DeleteAndStoreTripsInDb(tripsWithBusStops);
 
@@ -79,10 +80,9 @@ namespace DoCeluNaCzasWebApi.Services.UpdateService.Helpers
             DelayService.TripData = tripDataList.FirstOrDefault(x => x.Day.Date <= DateTime.Today);
             DelayService.SetChooseBusStopModelCollection(busStopDataModel, groupedJoinedTrips);
 
-            CacheService.CacheData(busStopDataModel, CacheKeys.BUS_STOP_DATA_MODEL);
-            CacheService.CacheData(groupedJoinedTrips, CacheKeys.GROUPED_JOINED_MODEL_LIST);
+            _documentStoreRepository.UpdateGroupedJoinedModels(groupedJoinedTrips);
 
-            _timeService.CacheLastUpdates(tripDataList.FirstOrDefault().LastUpdate, 
+            _timeService.CacheLastUpdates(tripDataList.FirstOrDefault().LastUpdate,
                                           busStopDataList.FirstOrDefault().LastUpdate,
                                           busStopDataList.FirstOrDefault().LastUpdate,
                                           stopInTripDataList.FirstOrDefault().LastUpdate,
@@ -91,7 +91,8 @@ namespace DoCeluNaCzasWebApi.Services.UpdateService.Helpers
 
         void DeleteAndStoreTripsInDb(List<TripsWithBusStops> tripsWithBusStops)
         {
-            _documentStoreRepository.DeleteTripsWithBusStops();
+            if (tripsWithBusStops.Count > 0)
+                _documentStoreRepository.DeleteTripsWithBusStops();
             _documentStoreRepository.Save(tripsWithBusStops);
         }
     }
